feat: rewrite in typescript
