+++ helper/replacer.ts
@@ -1,4 +1,4 @@
-export function HReplace(text, replaceKey, replaceValue) {
+export function HReplace(text: string, replaceKey: string[], replaceValue: string[]) {
   let result = text
 
   for (let i = 0; i < replaceKey.length; i++) {
