+++ package.json
@@ -3,21 +3,22 @@
   "version": "4.0.0",
   "description": "",
   "main": "server.js",
-  "type": "module",
+  "type": "commonjs",
   "scripts": {
-    "start:dev": "clear && nodemon server.js",
-    "start": "node server.js"
+    "build": "tsc",
+    "start": "npm run build && node dist/server.js",
+    "start:dev": "nodemon --exec ts-node -r server.ts"
   },
   "author": "",
-  "license": "ISC",
+  "license": "MIT",
   "dependencies": {
     "axios": "0.27.2",
     "dotenv": "16.0.1",
-    "express": "^4.18.2",
+    "express": "4.18.2",
     "fuzzy": "0.1.3",
     "grammy": "1.8.3",
     "grammy-scenes": "7.3.0",
-    "mongoose": "6.3.6",
+    "mongoose": "7.5.2",
     "node-cron": "3.0.2",
     "node-schedule": "2.1.0",
     "pdf-parse": "1.1.1"
@@ -28,5 +29,16 @@
     "#daily/*": "./daily/*.js",
     "#middlewares/*": "./middlewares/*.js",
     "#helper/*": "./helper/*.js"
+  },
+  "devDependencies": {
+    "@types/express": "4.17.17",
+    "@types/node": "20.6.2",
+    "@types/node-cron": "3.0.8",
+    "@types/node-schedule": "2.1.0",
+    "@types/pdf-parse": "1.1.1",
+    "nodemon": "3.0.1",
+    "ts-node": "10.9.1",
+    "tsconfig-paths": "4.2.0",
+    "typescript": "5.2.2"
   }
 }
